refactor: componentize UnifiedItemView for improved maintainability
- Extract UnifiedItemHeader, UnifiedItemActions, UnifiedItemChildren, and UnifiedItemName components
- Create unifiedItemDisplayUtils with reusable type display and event handler utilities
- Reduce code duplication by centralizing type-to-display mapping logic
- Improve readability through smaller, focused components with single responsibilities
- Maintain backward compatibility with existing component exports
