+++ src/sections/unified-item/components/UnifiedItemView.tsx
@@ -1,28 +1,19 @@
-import {
-  type Accessor,
-  createMemo,
-  createResource,
-  For,
-  type JSXElement,
-  Show,
-} from 'solid-js'
+import { type Accessor, createMemo, type JSXElement, Show } from 'solid-js'
 
-import { createSupabaseRecipeRepository } from '~/modules/diet/recipe/infrastructure/supabaseRecipeRepository'
-import { isRecipeUnifiedItemManuallyEdited } from '~/modules/diet/unified-item/domain/conversionUtils'
 import {
-  isGroupItem,
-  isRecipeItem,
+  asFoodItem,
+  isFoodItem,
   type UnifiedItem,
 } from '~/modules/diet/unified-item/schema/unifiedItemSchema'
 import {
   isFoodFavorite,
   setFoodAsFavorite,
 } from '~/modules/user/application/user'
-import { ContextMenu } from '~/sections/common/components/ContextMenu'
-import { CopyIcon } from '~/sections/common/components/icons/CopyIcon'
-import { MoreVertIcon } from '~/sections/common/components/icons/MoreVertIcon'
-import { TrashIcon } from '~/sections/common/components/icons/TrashIcon'
 import MacroNutrientsView from '~/sections/macro-nutrients/components/MacroNutrientsView'
+import { UnifiedItemActions } from '~/sections/unified-item/components/UnifiedItemActions'
+import { UnifiedItemChildren } from '~/sections/unified-item/components/UnifiedItemChildren'
+import { UnifiedItemHeader } from '~/sections/unified-item/components/UnifiedItemHeader'
+import { createEventHandler } from '~/sections/unified-item/utils/unifiedItemDisplayUtils'
 import { cn } from '~/shared/cn'
 import { createDebug } from '~/shared/utils/createDebug'
 import {
@@ -48,57 +39,6 @@ export type UnifiedItemViewProps = {
 
 export function UnifiedItemView(props: UnifiedItemViewProps) {
   const isInteractive = () => props.mode !== 'summary'
-  const hasChildren = () => {
-    const item = props.item()
-    return (
-      (isRecipeItem(item) || isGroupItem(item)) &&
-      Array.isArray(item.reference.children) &&
-      item.reference.children.length > 0
-    )
-  }
-
-  const getChildren = () => {
-    const item = props.item()
-    if (isRecipeItem(item) || isGroupItem(item)) {
-      return item.reference.children
-    }
-    return []
-  }
-
-  const handleMouseEvent = (callback?: () => void) => {
-    if (callback === undefined) {
-      return undefined
-    }
-    return (e: MouseEvent) => {
-      e.stopPropagation()
-      e.preventDefault()
-      callback()
-    }
-  }
-  const getHandlers = () => {
-    return {
-      onClick: handleMouseEvent(
-        props.handlers.onClick
-          ? () => props.handlers.onClick!(props.item())
-          : undefined,
-      ),
-      onEdit: handleMouseEvent(
-        props.handlers.onEdit
-          ? () => props.handlers.onEdit!(props.item())
-          : undefined,
-      ),
-      onCopy: handleMouseEvent(
-        props.handlers.onCopy
-          ? () => props.handlers.onCopy!(props.item())
-          : undefined,
-      ),
-      onDelete: handleMouseEvent(
-        props.handlers.onDelete
-          ? () => props.handlers.onDelete!(props.item())
-          : undefined,
-      ),
-    }
-  }
 
   return (
     <div
@@ -106,194 +46,27 @@ export function UnifiedItemView(props: UnifiedItemViewProps) {
         'block rounded-lg border border-gray-700 bg-gray-700 p-3 shadow hover:cursor-pointer hover:bg-gray-700',
         props.class,
       )}
-      onClick={(e) => getHandlers().onClick?.(e)}
+      onClick={(e: MouseEvent) => {
+        const handler = createEventHandler(props.handlers.onClick, props.item())
+        handler?.(e)
+      }}
     >
-      <div class="flex items-center">
-        <div class="flex flex-1  items-center">
-          <div class="flex-1">
-            {typeof props.header === 'function' ? props.header() : props.header}
-          </div>
-          <div class="">
-            {isInteractive() && (
-              <ContextMenu
-                trigger={
-                  <div class="text-3xl active:scale-105 hover:text-blue-200">
-                    <MoreVertIcon />
-                  </div>
-                }
-                class="ml-2"
-              >
-                <Show when={getHandlers().onEdit}>
-                  {(onEdit) => (
-                    <ContextMenu.Item
-                      class="text-left px-4 py-2 hover:bg-gray-700"
-                      onClick={onEdit()}
-                    >
-                      <div class="flex items-center gap-2">
-                        <span class="text-blue-500">✏️</span>
-                        <span>Editar</span>
-                      </div>
-                    </ContextMenu.Item>
-                  )}
-                </Show>
-                <Show when={getHandlers().onCopy}>
-                  {(onCopy) => (
-                    <ContextMenu.Item
-                      class="text-left px-4 py-2 hover:bg-gray-700"
-                      onClick={onCopy()}
-                    >
-                      <div class="flex items-center gap-2">
-                        <CopyIcon size={15} />
-                        <span>Copiar</span>
-                      </div>
-                    </ContextMenu.Item>
-                  )}
-                </Show>
-                <Show when={getHandlers().onDelete}>
-                  {(onDelete) => (
-                    <ContextMenu.Item
-                      class="text-left px-4 py-2 text-red-400 hover:bg-gray-700"
-                      onClick={onDelete()}
-                    >
-                      <div class="flex items-center gap-2">
-                        <span class="text-red-400">
-                          <TrashIcon size={15} />
-                        </span>
-                        <span class="text-red-400">Excluir</span>
-                      </div>
-                    </ContextMenu.Item>
-                  )}
-                </Show>
-              </ContextMenu>
-            )}
-          </div>
-        </div>
-      </div>
+      <UnifiedItemHeader item={props.item}>
+        {typeof props.header === 'function' ? props.header() : props.header}
+        {isInteractive() && (
+          <UnifiedItemActions item={props.item} handlers={props.handlers} />
+        )}
+      </UnifiedItemHeader>
+
       {typeof props.nutritionalInfo === 'function'
         ? props.nutritionalInfo()
         : props.nutritionalInfo}
-      <Show when={hasChildren()}>
-        <div class="mt-2 ml-4 space-y-1">
-          <For each={getChildren()}>
-            {(child) => (
-              <div class="text-sm text-gray-300 flex justify-between">
-                <span>
-                  {child.name} ({child.quantity}g)
-                </span>
-                <span class="text-gray-400">
-                  {calcUnifiedItemCalories(child).toFixed(0)}kcal
-                </span>
-              </div>
-            )}
-          </For>
-        </div>
-      </Show>
+
+      <UnifiedItemChildren item={props.item} />
+
       <Show when={props.item().reference.type === 'food'}>
         <UnifiedItemFavorite foodId={props.item().id} />
       </Show>
-    </div>
-  )
-}
-
-export function UnifiedItemName(props: { item: Accessor<UnifiedItem> }) {
-  const recipeRepository = createSupabaseRecipeRepository()
-
-  // Create a resource to fetch recipe data when needed
-  const [originalRecipe] = createResource(
-    () => {
-      const item = props.item()
-      return isRecipeItem(item) ? item.reference.id : null
-    },
-    async (recipeId: number) => {
-      try {
-        return await recipeRepository.fetchRecipeById(recipeId)
-      } catch (error) {
-        console.warn('Failed to fetch recipe for comparison:', error)
-        return null
-      }
-    },
-  )
-
-  // Check if the recipe was manually edited
-  const isManuallyEdited = createMemo(() => {
-    const item = props.item()
-    const recipe = originalRecipe()
-
-    if (
-      !isRecipeItem(item) ||
-      recipe === null ||
-      recipe === undefined ||
-      originalRecipe.loading
-    ) {
-      return false
-    }
-
-    return isRecipeUnifiedItemManuallyEdited(item, recipe)
-  })
-
-  const nameColor = () => {
-    const item = props.item()
-
-    switch (item.reference.type) {
-      case 'food':
-        return 'text-white'
-      case 'recipe':
-        return 'text-yellow-200'
-      case 'group':
-        return 'text-green-200'
-      default:
-        return 'text-gray-400'
-    }
-  }
-
-  const typeIndicator = () => {
-    const item = props.item()
-    switch (item.reference.type) {
-      case 'food':
-        return '🍽️'
-      case 'recipe':
-        return '📖'
-      case 'group':
-        return '📦'
-      default:
-        return '❓'
-    }
-  }
-
-  const getTypeText = () => {
-    const item = props.item()
-    switch (item.reference.type) {
-      case 'food':
-        return 'alimento'
-      case 'recipe':
-        return 'receita'
-      case 'group':
-        return 'grupo'
-      default:
-        return 'desconhecido'
-    }
-  }
-
-  const warningIndicator = () => {
-    return isManuallyEdited() ? '⚠️' : ''
-  }
-
-  return (
-    <div class="">
-      <h5 class={`mb-2 text-lg font-bold tracking-tight ${nameColor()}`}>
-        <span class="mr-2 cursor-help" title={getTypeText()}>
-          {typeIndicator()}
-        </span>
-        {props.item().name}
-        <Show when={warningIndicator()}>
-          <span
-            class="ml-1 text-yellow-500"
-            title="Receita editada pontualmente"
-          >
-            {warningIndicator()}
-          </span>
-        </Show>
-      </h5>
     </div>
   )
 }
@@ -337,3 +110,6 @@ export function UnifiedItemFavorite(props: { foodId: number }) {
     </div>
   )
 }
+
+// Re-export for backward compatibility
+export { UnifiedItemName } from '~/sections/unified-item/components/UnifiedItemName'
